refactor: MenuDrawer under MenuBar
