+++ components/MenuBar/MenuDrawer/SettingsMenu.tsx
@@ -12,9 +12,9 @@ import SettingsIcon from '@mui/icons-material/Settings'
 import SystemUpdateIcon from '@mui/icons-material/SystemUpdate'
 import RestoreIcon from '@mui/icons-material/Restore'
 import { useDispatch } from 'react-redux'
-import { layoutActions } from '../../redux/slices/layoutSlice'
+import { layoutActions } from '../../../redux/slices/layoutSlice'
 import { useSelector } from 'react-redux'
-import { RootState } from '../../redux/store'
+import { RootState } from '../../../redux/store'
 import { MenuButton, MenuButtonGrouper } from './common'
 import { Dispatch, SetStateAction, useState } from 'react'
