+++ components/MenuBar/index.tsx
@@ -9,7 +9,7 @@ import MenuIcon from '@mui/icons-material/Menu'
 import LightModeIcon from '@mui/icons-material/LightMode'
 import DarkModeIcon from '@mui/icons-material/DarkMode'
 import { useState } from "react"
-import MenuDrawer from "../MenuDrawer"
+import MenuDrawer from "./MenuDrawer"
 import { useDispatch } from "react-redux"
 import { themeActions } from "../../redux/slices/themeSlice"
 import { APP_NAME, VERSION } from "../../constants"
